==> analysis/results/2022_05_16_cross_continental_endemic/2022_05_16_batch_0/4_surveillance_merge.R <==
managementDfPaths = list.files("./output", "resultsDfSummary_.*.rds", full.names = TRUE)

stackedDfList = list()
for(i in managementDfPaths){
    stackedDfList[[i]] = readRDS(i)
}

x = dplyr::bind_rows(stackedDfList)

saveRDS(x, "./output/resultsDfSummary.rds")
